@@ -4,53 +4,30 @@
   "description": "",
   "main": "gulpfile.babel.js",
   "dependencies": {
-    "babel": "^5.6.23",
-    "babel-core": "^5.8.20",
-    "babelify": "^6.1.3",
-    "browser-sync": "^2.7.13",
-    "express": "^4.13.1",
-    "flux": "^2.0.3",
-    "gulp": "^3.9.0",
-    "gulp-babel": "^5.1.0",
-    "gulp-streamify": "^0.0.5",
-    "gulp-uglify": "^1.2.0",
-    "jade": "^1.11.0",
-    "keymirror": "^0.1.1",
-    "lodash": "^3.10.0",
-    "object-assign": "^3.0.0",
-    "react": "^0.13.3",
-    "react-router": "^0.13.3",
-    "react-tools": "^0.13.3",
-    "run-parallel": "^1.1.2",
-    "vinyl-source-stream": "^1.1.0",
-    "when": "^3.7.3"
+    "react": "^15.1.0",
+    "react-dom": "^15.1.0",
+    "react-router": "^2.4.1"
   },
   "devDependencies": {
-    "autoprefixer-core": "^5.2.1",
-    "babel-eslint": "^3.1.23",
-    "babel-jest": "^5.3.0",
-    "babel-loader": "^5.3.2",
+    "babel-core": "^6.9.1",
+    "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-react-hmre": "^1.1.1",
+    "babel-preset-stage-0": "^6.5.0",
     "browserify": "^11.0.1",
-    "css-loader": "^0.17.0",
-    "eslint": "^1.0.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-react": "^2.7.1",
-    "gulp-sass": "^2.0.4",
-    "gulp-util": "^3.0.6",
-    "jest-cli": "^0.4.15",
-    "less-loader": "^2.2.0",
-    "minimist": "^1.1.1",
-    "node-libs-browser": "^0.5.2",
-    "node-sass": "^3.3.2",
-    "postcss-loader": "^0.6.0",
-    "reactify": "^1.1.1",
-    "run-sequence": "^1.1.1",
-    "sass-loader": "^2.0.1",
-    "source-map-support": "^0.3.2",
-    "style-loader": "^0.12.3",
-    "webpack": "^1.10.1"
+    "css-loader": "^0.23.1",
+    "file-loader": "^0.8.5",
+    "hjs-webpack": "^8.2.0",
+    "node-sass": "^3.7.0",
+    "postcss-loader": "^0.9.1",
+    "sass-loader": "^3.2.0",
+    "style-loader": "^0.13.1",
+    "url-loader": "^0.5.7",
+    "webpack": "^1.13.1"
   },
   "scripts": {
+    "start": "NODE_ENV=development ./node_modules/.bin/hjs-dev-server",
     "test": "jest"
   },
   "jest": {
